move android controller methods into their own file
